refactor(storage): complete mechanical composition; restore build
- Move 169 types, 78 methods, and constants from internal/storage/models_*.go
  to internal/types/models_migrated.go and internal/types/types.go
- Add type aliases and const re-exports in internal/storage/interfaces.go
  to preserve the public API (storage.SignalFilter, storage.BillingSystemType,
  storage.MFAMode*, storage.ReferralStatus*, storage.CommissionStatus*, etc.)
- Add Function = RegistryFunction alias for backwards compatibility
- Add Function relation to FunctionFollow struct (used by follow handlers)
- Add DeployKeyListResponse, FunctionWebhookDeliveryListResponse, and
  FunctionWebhookTestRequest types to support their handlers
- Add StripeSyncStatusPending, StripeSyncStatusIgnored, MembershipStatus*
  constants referenced by webhooks and auth handlers
- Update callers to pass context.Context to repo methods (autoreadme,
  modelresolver, flyembed-backfill, setup, billing/tenants)
- Restore env helpers in internal/api/middleware/log_sampler.go
  (getEnvFloatFromEnv, getEnvInt, getEnvDuration)
- Fix internal/fom/collector.go: replace log.Error (stdlib has no such method)
  with log.Printf; fix c.buffer channel-vs-slice bug to use append+mutex
- Fix internal/fom/physics/hint_engine.go: rename shadowed context param
  to execCtx, remove unused time import in engine.go
- Fix internal/api/handlers/admin/billing.go: filter.Status is *string
- Fix internal/api/handlers/admin/tenants.go: h.Rate is int, not int64
- Fix internal/api/handlers/deploykeys/handler.go: cast int64 total to int
- Add cmd/fly/commands/client.go with vaultAPIClient stub (Get/Post/Delete/Patch
  return helpful error pointing to functionfly/ff-cli) so orphan vault CLI
  commands still compile
- Replace cmd/ffly/main.go with redirect-to-ff-cli stub
- Add new internal/api/middleware files (error_normalizer, rate_limiter_*,
  log_sampler, timeout) and internal/auth/hibp.go
- Add internal/dre/cert/anchoring_signer.go
- Add internal/storage/batch_cleanup.go and cleanup_coordinator.go
- Add internal/storage/mock_Repository.go (mockery generated)
- Add migrations/20260614195400_audit_integrity_hash_chain

go build ./internal/... ./cmd/... now exits 0.
The only remaining go build ./... error is the pre-existing
examples/wasm cgo issue (unrelated).

+++ internal/api/handlers/admin/tenants.go
@@ -436,7 +436,7 @@ func (h *Handler) HandleTenantMetrics(w http.ResponseWriter, r *http.Request) {
 		logrus.WithError(err).WithField("tenant_id", tenantID).Warn("Failed to get execution rate for tenant metrics")
 	}
 
-	executions24h := int64(0)
+	executions24h := 0
 	for _, h := range execRate24h {
 		executions24h += h.Rate
 	}
